Add files via upload
This is a PowerShell function named "Remove-IEIconFromTaskbar" that removes the Internet Explorer icon from the taskbar configuration on a remote computer.
The function takes one parameter, $ComputerName, which specifies the name of the remote computer.The function first defines an XML namespace to search for the target element,
then gets all user profile folders, including the Default User profile folder. For each user profile, it defines the file path of LayoutModification.xml and checks if the file exists.
If the file exists, it loads the XML content from the file and finds the target element to remove using an XPath query that searches for a specific DesktopApp element that represents 
the Internet Explorer icon on the taskbar. If the target element is found, it removes it from the XML content and saves the modified XML content back to the file. The function also writes
 output to the console to indicate whether the icon was found and removed or not, and for which user profile.To run the function, the user is prompted to enter the remote computer name, 
 and then the function is invoked using the specified computer name as the parameter value. The Invoke-Command cmdlet is used to run the script block on the remote computer